fix: cast packed PayloadHeader::offset fields to avoid reference binding
GCC/clang reject passing packed struct fields directly to std::format
when it tries to bind them by reference. Cast the uint32_t offset field
to (uint32_t) explicitly at all four log call sites in
download_burst_continue() — the same pattern used for seq_number in the
previous commit.

Fixes: "error: cannot bind packed field '...::offset' to 'unsigned int&'"
at lines 532, 541, 574, and 612 of mavlink_ftp_client.cpp.

+++ cpp/src/mavsdk/core/mavlink_ftp_client.cpp
@@ -529,7 +529,7 @@ bool MavlinkFtpClient::download_burst_continue(
 
     } else if (payload->req_opcode == CMD_BURST_READ_FILE) {
         if (_debugging) {
-            LogDebug("Burst download continue, at: {} write: {}", payload->offset, payload->size);
+            LogDebug("Burst download continue, at: {} write: {}", (uint32_t)payload->offset, (int)payload->size);
         }
 
         if (payload->offset != item.current_offset) {
@@ -538,7 +538,7 @@ bool MavlinkFtpClient::download_burst_continue(
                 // it.
                 LogWarn(
                     "Got payload offset: {}, next offset: {}",
-                    payload->offset,
+                    (uint32_t)payload->offset,
                     item.current_offset);
                 return false;
             }
@@ -571,7 +571,7 @@ bool MavlinkFtpClient::download_burst_continue(
         item.current_offset = payload->offset + payload->size;
 
         if (_debugging) {
-            LogDebug("Received {} to {}", payload->offset, payload->size + payload->offset);
+            LogDebug("Received {} to {}", (uint32_t)payload->offset, payload->size + payload->offset);
         }
 
         if (payload->size + payload->offset >= item.file_size) {
@@ -609,8 +609,8 @@ bool MavlinkFtpClient::download_burst_continue(
         if (_debugging) {
             LogWarn(
                 "Burst download continue missing pieces, write at {} for {}",
-                payload->offset,
-                payload->size);
+                (uint32_t)payload->offset,
+                (int)payload->size);
         }
 
         item.ofstream.seekp(payload->offset);
